Mauro: 4.2 Grupos de trabajo. Falta eliminar seleccionados (Pag 53) [IT'S DONE]

+++ apps/frontend/modules/grupos_de_trabajo/actions/actions.class.php
@@ -71,19 +71,41 @@ class grupos_de_trabajoActions extends sfActions
 
   public function executeDelete(sfWebRequest $request)
   {
-    $request->checkCSRFProtection();
-
-    $this->forward404Unless($grupos_de_trabajo = Doctrine::getTable('GrupoTrabajo')->find($request->getParameter('id')), sprintf('Object grupos_de_trabajo does not exist (%s).', $request->getParameter('id')));
-    
-    sfLoader::loadHelpers('Security'); // para usar el helper
-		if (!validate_action('baja')) $this->redirect('seguridad/restringuido');
-    
-    $grupos_de_trabajo->delete();
-
-		$this->getUser()->setFlash('notice', "El registro ha sido eliminado del sistema");
-    $this->redirect('grupos_de_trabajo/index');
+  	$this->processSelectedRecords($request, 'baja');
+ 
   }
 
+  protected function processSelectedRecords(sfWebRequest $request, $accion)
+  {
+  	$toProcess = $request->getParameter('id');
+  	
+  	if (!empty($toProcess)) {
+  		$request->checkCSRFProtection();
+  		
+  		$IDs = is_array($toProcess) ? $toProcess : array($toProcess);
+  		
+  		foreach ($IDs as $id) {
+  			    $this->forward404Unless($grupos_de_trabajo = Doctrine::getTable('GrupoTrabajo')->find($id), sprintf('Object documentacion_grupo does not exist (%s).', $id));
+  			
+  			    sfLoader::loadHelpers('Security');
+				if (!validate_action($accion)) $this->redirect('seguridad/restringuido');
+
+				if ($accion == 'publicar') {
+//						$documentacion_grupo->setEstado('publicado');
+//						$documentacion_grupo->save();
+//			
+//						ServiceNotificacion::send('creacion', 'Grupo', $documentacion_grupo->getId(), $documentacion_grupo->getNombre(),'',$documentacion_grupo->getGrupoTrabajoId());	
+				} else {
+					
+				    $grupos_de_trabajo->delete();
+				
+					$this->getUser()->setFlash('notice', "El registro ha sido eliminado del sistema");
+				}		
+  		}
+  	}
+  $this->redirect('grupos_de_trabajo/index');
+  }
+  
   protected function processForm(sfWebRequest $request, sfForm $form, $accion='')
   {
     $form->bind($request->getParameter($form->getName()));
